Refactor into association a module
