+++ classes/app.rb
@@ -1,12 +1,9 @@
 require_relative './app_helpers'
-require_relative 'game'
-require_relative 'genre'
-require_relative 'source'
-require_relative 'label'
-require_relative 'author'
+require_relative 'associations'
 require 'date'
 class App
   include AppHelpers
+  include Associations
   def initialize
     @books = []
     @games = []
@@ -59,16 +56,6 @@ class App
 
   private
 
-  def list_items(list)
-    list.each_with_index do |item, i|
-      puts %(#{i + 1}. #{item.label.title}
-        Source: #{item.source.name}
-        Author: #{item.author.first_name} #{item.author.last_name}
-        Genre: #{item.genre.name}
-      )
-    end
-  end
-
   def add_association(item)
     add_associate(label, @labels, item) { |label| item.add_label label }
     add_associate(author, @authors, item) { |author| item.add_author author }
@@ -76,41 +63,13 @@ class App
     add_associate(genre, @genres, item) { |genre| item.add_genre genre }
   end
 
-  def add_associate(associate, list, item)
-    associate.add_item item
-    yield associate
-    list << associate
-  end
-
-  def label
-    puts 'Lets add the Label'
-    print 'title: '
-    title = gets.chomp
-    print 'color: '
-    color = gets.chomp
-    Label.new(title: title, color: color)
-  end
-
-  def source
-    puts 'Lets add the Source'
-    print 'Source name: '
-    name = gets.chomp
-    Source.new(name: name)
-  end
-
-  def genre
-    puts 'Lets add the Genre'
-    print 'Genre name: '
-    name = gets.chomp
-    Genre.new(name: name)
-  end
-
-  def author
-    puts "Let's add the Author"
-    print 'First name: '
-    first_name = gets.chomp
-    print 'Last name: '
-    last_name = gets.chomp
-    Author.new(first_name: first_name, last_name: last_name)
+  def list_items(list)
+    list.each_with_index do |item, i|
+      puts %(#{i + 1}. #{item.label.title}
+      Source: #{item.source.name}
+      Author: #{item.author.first_name} #{item.author.last_name}
+      Genre: #{item.genre.name}
+      )
+    end
   end
 end
